requirements.txt
# SPARC Voice Agent Backend - Dependencies

# Web framework
fastapi==0.135.1
uvicorn==0.42.0

# OpenAI - LLM + embeddings
openai==2.29.0

# Pinecone - vector store
pinecone==8.1.0

# Document processing
python-docx==1.2.0

# Environment
python-dotenv==1.2.2

# HTTP client (for Vapi API setup)
requests==2.32.5
